Add types for event callback functions
Hi! I just want to say, thank you for creating Emitter. It's simple and easy to use.

So, currently, callback functions args for an event are of type `any`. Here I implemented the same technique TS uses for DOM events types. Now when you pass the first argument, the second argument's type is inferred.

As you can see, we still need to add types for the `connack` and `error` args. I'm hoping you can help complete this or at least point me in the right direction.